fix bug when not specifying box constraints
